requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python-headless
python-multipart 
requests 
python-telegram-bot
python-dotenv
httpx  # <-- ADD THIS LINE